+++ pkg/analysis_server/lib/src/lsp/mapping.dart
@@ -222,9 +222,8 @@ WorkspaceEdit createRenameEdit(
 
 /// Creates a [lsp.WorkspaceEdit] from a [server.SourceChange].
 ///
-/// Can return experimental [lsp.LegacySnippetTextEdit]s if the following are true:
-/// - the client has indicated support for in the experimental section of their
-///   client capabilities, and
+/// Can include [lsp.SnippetTextEdit]s or [lsp.LegacySnippetTextEdit] if the
+/// client has indicated support for one or the other, and
 /// - [allowSnippets] is true, and
 /// - [change] contains only a single edit to the single file [filePath]
 /// - [lineInfo] is provided (which should be for the single edited file)
@@ -251,7 +250,8 @@ lsp.WorkspaceEdit createWorkspaceEdit(
   // existing file with a single edit and that there is either a selection or a
   // linked edit group (otherwise there's no value in snippets).
   if (!allowSnippets ||
-      !clientCapabilities.legacySnippetTextEdit ||
+      (!clientCapabilities.legacySnippetTextEdit &&
+          !clientCapabilities.snippetTextEdit) ||
       !clientCapabilities.documentChanges ||
       filePath == null ||
       lineInfo == null ||
@@ -271,28 +271,48 @@ lsp.WorkspaceEdit createWorkspaceEdit(
   }
 
   var fileEdit = change.edits.single;
-  var snippetEdits = toLegacySnippetTextEdits(
-    fileEdit.file,
-    fileEdit,
-    change.linkedEditGroups,
-    lineInfo,
-    selectionOffset: change.selection?.offset,
-    selectionLength: change.selectionLength,
-  );
+  var snippetEdits = clientCapabilities.snippetTextEdit
+      ? toSnippetTextEdits(
+              fileEdit.file,
+              fileEdit,
+              change.linkedEditGroups,
+              lineInfo,
+              selectionOffset: change.selection?.offset,
+              selectionLength: change.selectionLength,
+            )
+            .map(
+              Either4<
+                    AnnotatedTextEdit,
+                    LegacySnippetTextEdit,
+                    SnippetTextEdit,
+                    TextEdit
+                  >
+                  .t3,
+            )
+            .toList()
+      : toLegacySnippetTextEdits(
+              fileEdit.file,
+              fileEdit,
+              change.linkedEditGroups,
+              lineInfo,
+              selectionOffset: change.selection?.offset,
+              selectionLength: change.selectionLength,
+            )
+            .map(
+              Either4<
+                    AnnotatedTextEdit,
+                    LegacySnippetTextEdit,
+                    SnippetTextEdit,
+                    TextEdit
+                  >
+                  .t2,
+            )
+            .toList();
 
   // Compile the edits into a TextDocumentEdit for this file.
   var textDocumentEdit = lsp.TextDocumentEdit(
     textDocument: analysisServer.getVersionedDocumentIdentifier(fileEdit.file),
-    edits: snippetEdits
-        .map(
-          (e) =>
-              Either3<
-                lsp.AnnotatedTextEdit,
-                lsp.LegacySnippetTextEdit,
-                lsp.TextEdit
-              >.t2(e),
-        )
-        .toList(),
+    edits: snippetEdits,
   );
 
   // Convert to the union that documentChanges require.
@@ -934,6 +954,61 @@ ChangeAnnotation? recordEditAnnotation(
 ///       0 -> 9999999 -       0 -> 9 999 999
 String relevanceToSortText(int relevance) =>
     (sortTextMaxValue - relevance).toString();
+
+/// Creates a [lsp.SnippetTextEdit] for a set of edits using Linked Edit
+/// Groups.
+///
+/// Edit groups offsets are based on the entire content being modified after all
+/// edits, so [editOffset] must to take into account both the offset of the edit
+/// _and_ any delta from edits prior to this one in the file.
+///
+/// [selectionOffset] is also absolute and assumes `edit.replacement` will be
+/// inserted at [editOffset].
+lsp.SnippetTextEdit snippetTextEditFromEditGroups(
+  String filePath,
+  server.LineInfo lineInfo,
+  server.SourceEdit edit, {
+  required List<server.LinkedEditGroup> editGroups,
+  required int editOffset,
+  required int? selectionOffset,
+  required int? selectionLength,
+}) {
+  return lsp.SnippetTextEdit(
+    range: toRange(lineInfo, edit.offset, edit.length),
+    snippet: StringValue(
+      value: buildSnippetStringForEditGroups(
+        edit.replacement,
+        filePath: filePath,
+        editGroups: editGroups,
+        editOffset: editOffset,
+        selectionOffset: selectionOffset,
+        selectionLength: selectionLength,
+      ),
+    ),
+  );
+}
+
+/// Creates a [lsp.SnippetTextEdit] for an edit with a selection
+/// placeholder.
+///
+/// [selectionOffsetRelative] is relative to (and therefore must be within) the
+/// edit.
+lsp.SnippetTextEdit snippetTextEditWithSelection(
+  server.LineInfo lineInfo,
+  server.SourceEdit edit, {
+  required int selectionOffsetRelative,
+  int? selectionLength,
+}) {
+  return lsp.SnippetTextEdit(
+    range: toRange(lineInfo, edit.offset, edit.length),
+    snippet: StringValue(
+      value: buildSnippetStringWithTabStops(edit.replacement, [
+        selectionOffsetRelative,
+        selectionLength ?? 0,
+      ]),
+    ),
+  );
+}
 
 lsp.CompletionItem snippetToCompletionItem(
   lsp.LspAnalysisServer server,
@@ -1638,6 +1713,44 @@ lsp.SignatureHelp toSignatureHelp(
   );
 }
 
+List<lsp.SnippetTextEdit> toSnippetTextEdits(
+  String filePath,
+  server.SourceFileEdit change,
+  List<server.LinkedEditGroup> editGroups,
+  LineInfo lineInfo, {
+  required int? selectionOffset,
+  required int? selectionLength,
+}) {
+  var snippetEdits = <lsp.SnippetTextEdit>[];
+
+  // Edit groups offsets are based on the document after the edits are applied.
+  // This means we must compute an offset delta for each edit that takes into
+  // account all edits that might be made before it in the document (which are
+  // after it in the edits). To do this, reverse the list when computing the
+  // offsets, but reverse them back to the original list order when returning so
+  // that we do not apply them incorrectly in tests (where we will apply them
+  // in-sequence).
+
+  var offsetDelta = 0;
+  for (var edit in change.edits.reversed) {
+    snippetEdits.add(
+      snippetTextEditFromEditGroups(
+        filePath,
+        lineInfo,
+        edit,
+        editGroups: editGroups,
+        editOffset: edit.offset + offsetDelta,
+        selectionOffset: selectionOffset,
+        selectionLength: selectionLength,
+      ),
+    );
+
+    offsetDelta += edit.replacement.length - edit.length;
+  }
+
+  return snippetEdits.reversed.toList();
+}
+
 ErrorOr<server.SourceRange> toSourceRange(
   server.LineInfo lineInfo,
   Range range,
@@ -1691,7 +1804,9 @@ lsp.TextDocumentEdit toTextDocumentEdit(
   );
 }
 
-Either3<lsp.AnnotatedTextEdit, lsp.LegacySnippetTextEdit, lsp.TextEdit>
+/// Converts a [server.SourceEdit] into an LSP edit depending on the clients
+/// capabilities for snippets and/or annotated changes.
+Either4<AnnotatedTextEdit, LegacySnippetTextEdit, SnippetTextEdit, TextEdit>
 toTextDocumentEditEdit(
   LspClientCapabilities capabilities,
   server.LineInfo lineInfo,
@@ -1700,38 +1815,37 @@ toTextDocumentEditEdit(
   int? selectionLength,
   lsp.ChangeAnnotationIdentifier? annotationIdentifier,
 }) {
-  if (annotationIdentifier != null) {
-    return Either3<
-      lsp.AnnotatedTextEdit,
-      lsp.LegacySnippetTextEdit,
-      lsp.TextEdit
-    >.t1(
+  var prefersSnippet = selectionOffsetRelative != null;
+
+  if (prefersSnippet && capabilities.snippetTextEdit) {
+    return .t3(
+      snippetTextEditWithSelection(
+        lineInfo,
+        edit,
+        selectionOffsetRelative: selectionOffsetRelative,
+        selectionLength: selectionLength,
+      ),
+    );
+  } else if (prefersSnippet && capabilities.legacySnippetTextEdit) {
+    return .t2(
+      legacySnippetTextEditWithSelection(
+        lineInfo,
+        edit,
+        selectionOffsetRelative: selectionOffsetRelative,
+        selectionLength: selectionLength,
+      ),
+    );
+  } else if (annotationIdentifier != null) {
+    return .t1(
       lsp.AnnotatedTextEdit(
         annotationId: annotationIdentifier,
         range: toRange(lineInfo, edit.offset, edit.length),
         newText: edit.replacement,
       ),
     );
+  } else {
+    return .t4(toTextEdit(lineInfo, edit));
   }
-  if (!capabilities.legacySnippetTextEdit || selectionOffsetRelative == null) {
-    return Either3<
-      lsp.AnnotatedTextEdit,
-      lsp.LegacySnippetTextEdit,
-      lsp.TextEdit
-    >.t3(toTextEdit(lineInfo, edit));
-  }
-  return Either3<
-    lsp.AnnotatedTextEdit,
-    lsp.LegacySnippetTextEdit,
-    lsp.TextEdit
-  >.t2(
-    legacySnippetTextEditWithSelection(
-      lineInfo,
-      edit,
-      selectionOffsetRelative: selectionOffsetRelative,
-      selectionLength: selectionLength,
-    ),
-  );
 }
 
 lsp.TextEdit toTextEdit(