+++ pkg/analysis_server/test/shared/shared_code_actions_assists_tests.dart
@@ -16,6 +16,7 @@ import 'package:test/test.dart';
 import '../lsp/code_actions_mixin.dart';
 import '../lsp/request_helpers_mixin.dart';
 import '../lsp/server_abstract.dart';
+import '../utils/lsp_protocol_extensions.dart';
 import '../utils/test_code_extensions.dart';
 import 'shared_test_interface.dart';
 
@@ -242,17 +243,9 @@ void f() {
 
     // Also ensure there was a single edit that was correctly marked
     // as a LegacySnippetTextEdit.
-    var textEdits = extractTextDocumentEdits(verifier.edit.documentChanges!)
-        .expand((tde) => tde.edits)
-        .map(
-          (edit) => edit.map(
-            (e) =>
-                throw 'Expected LegacySnippetTextEdit, got AnnotatedTextEdit',
-            (e) => e,
-            (e) => throw 'Expected LegacySnippetTextEdit, got TextEdit',
-          ),
-        )
-        .toList();
+    var textEdits = extractTextDocumentEdits(
+      verifier.edit.documentChanges!,
+    ).expand((tde) => tde.edits).map(_extractLegacySnippetTextEdit).toList();
     expect(textEdits, hasLength(1));
     expect(textEdits.first.insertTextFormat, equals(InsertTextFormat.Snippet));
   }
@@ -357,17 +350,9 @@ build() {
 
     // Also ensure there was a single edit that was correctly marked
     // as a LegacySnippetTextEdit.
-    var textEdits = extractTextDocumentEdits(verifier.edit.documentChanges!)
-        .expand((tde) => tde.edits)
-        .map(
-          (edit) => edit.map(
-            (e) =>
-                throw 'Expected LegacySnippetTextEdit, got AnnotatedTextEdit',
-            (e) => e,
-            (e) => throw 'Expected LegacySnippetTextEdit, got TextEdit',
-          ),
-        )
-        .toList();
+    var textEdits = extractTextDocumentEdits(
+      verifier.edit.documentChanges!,
+    ).expand((tde) => tde.edits).map(_extractLegacySnippetTextEdit).toList();
     expect(textEdits, hasLength(1));
     expect(textEdits.first.insertTextFormat, equals(InsertTextFormat.Snippet));
   }
@@ -402,11 +387,159 @@ Future? f;
     expect(codeActions, isEmpty);
   }
 
-  Future<void> test_snippetTextEdits_unsupported() async {
-    // This tests experimental support for including Snippets in TextEdits
-    // is not active when the client capabilities do not advertise support for it.
-    // https://github.com/rust-analyzer/rust-analyzer/blob/b35559a2460e7f0b2b79a7029db0c5d4e0acdb44/docs/dev/lsp-extensions.md#snippet-textedit
+  Future<void> test_snippetTextEdit_editGroupsAndSelection() async {
+    const content = '''
+void f() {
+  [!print(0);!]
+}
+''';
 
+    const expectedContent = r'''
+void f() {
+  if (${1:condition}) {
+    print(0);
+  }$0
+}
+''';
+
+    setSnippetTextEditSupport();
+    var verifier = await verifyCodeActionLiteralEdits(
+      content,
+      expectedContent,
+      kind: CodeActionKind('refactor.surround.if'),
+      title: "Surround with 'if'",
+    );
+
+    // Also ensure there was a single edit that was correctly marked
+    // as a SnippetTextEdit.
+    var textEdits = extractTextDocumentEdits(
+      verifier.edit.documentChanges!,
+    ).expand((tde) => tde.edits).toList();
+    expect(textEdits, hasLength(1));
+    var edit = textEdits.single.map((e) => e, (e) => e, (e) => e, (e) => e);
+    expect(edit, isA<SnippetTextEdit>());
+  }
+
+  Future<void> test_snippetTextEdits_multiEditGroup() async {
+    // As test_snippetTextEdits_singleEditGroup, but uses an assist that
+    // produces multiple linked edit groups.
+
+    const content = '''
+import 'package:flutter/widgets.dart';
+build() {
+  return Container(
+    child: Ro^w(
+      children: [
+        Text('111'),
+        Text('222'),
+        Container(),
+      ],
+    ),
+  );
+}
+''';
+
+    const expectedContent = r'''
+import 'package:flutter/widgets.dart';
+build() {
+  return Container(
+    child: ${1:widget}(
+      ${2:child}: Row(
+        children: [
+          Text('111'),
+          Text('222'),
+          Container(),
+        ],
+      ),
+    ),
+  );
+}
+''';
+
+    setSnippetTextEditSupport();
+    await verifyCodeActionLiteralEdits(
+      content,
+      expectedContent,
+      kind: CodeActionKind('refactor.flutter.wrap.generic'),
+      title: 'Wrap with widget...',
+    );
+  }
+
+  Future<void> test_snippetTextEdits_prefersStandardOverCustom() async {
+    // Verify that if we support both the old legacy snippet text edits and
+    // the new standard ones, that we use the new standard ones.
+    //
+    // Do this by just enabling legacy support and then triggering the
+    // new standard test.
+    setLegacySnippetTextEditSupport();
+    await test_snippetTextEdits_singleEditGroup();
+  }
+
+  Future<void> test_snippetTextEdits_singleEditGroup() async {
+    // This tests LSP v3.18 support for including Snippets in TextEdits.
+    //
+    // This allows setting the cursor position/selection in TextEdits included
+    // in CodeActions, for example Flutter's "Wrap with widget" assist that
+    // should select the text "widget".
+
+    const content = '''
+import 'package:flutter/widgets.dart';
+build() {
+  return Container(
+    child: Row(
+      children: [^
+        Text('111'),
+        Text('222'),
+        Container(),
+      ],
+    ),
+  );
+}
+''';
+
+    // For testing, the snippet will be inserted literally into the text, as
+    // this requires some magic on the client. The expected text should
+    // therefore contain the snippets in the standard format.
+    const expectedContent = r'''
+import 'package:flutter/widgets.dart';
+build() {
+  return Container(
+    child: Row(
+      children: [
+        ${0:widget}(
+          children: [
+            Text('111'),
+            Text('222'),
+            Container(),
+          ],
+        ),
+      ],
+    ),
+  );
+}
+''';
+
+    setSnippetTextEditSupport();
+    var verifier = await verifyCodeActionLiteralEdits(
+      content,
+      expectedContent,
+      kind: CodeActionKind('refactor.flutter.wrap.generic'),
+      title: 'Wrap with widget...',
+    );
+
+    // Also ensure there was a single edit that was correctly marked
+    // as a SnippetTextEdit.
+    var textEdits = extractTextDocumentEdits(
+      verifier.edit.documentChanges!,
+    ).expand((tde) => tde.edits).toList();
+    expect(textEdits, hasLength(1));
+    var edit = textEdits.single.map((e) => e, (e) => e, (e) => e, (e) => e);
+    expect(edit, isA<SnippetTextEdit>());
+  }
+
+  /// Ensure we don't get either kind of snippet text edits when we don't
+  /// support either.
+  Future<void> test_snippetTextEdits_unsupported() async {
     const content = '''
 import 'package:flutter/widgets.dart';
 build() {
@@ -431,16 +564,14 @@ build() {
     // Extract just TextDocumentEdits, create/rename/delete are not relevant.
     var edit = assist.edit!;
     var textDocumentEdits = extractTextDocumentEdits(edit.documentChanges!);
-    var textEdits = textDocumentEdits
-        .expand((tde) => tde.edits)
-        .map((edit) => edit.map((e) => e, (e) => e, (e) => e))
-        .toList();
 
-    // Ensure the edit does _not_ have a format of Snippet, nor does it include
+    // Ensure the edits do _not_ have any format of Snippet, nor include
     // any $ characters that would indicate snippet text.
-    for (var edit in textEdits) {
-      expect(edit, isNot(TypeMatcher<LegacySnippetTextEdit>()));
-      expect(edit.newText, isNot(contains(r'$')));
+    for (var editUnion in textDocumentEdits.expand((tde) => tde.edits)) {
+      var edit = editUnion.map((e) => e, (e) => e, (e) => e, (e) => e);
+      expect(edit, isNot(isA<LegacySnippetTextEdit>()));
+      expect(edit, isNot(isA<SnippetTextEdit>()));
+      expect(editUnion.extractTextEdit().newText, isNot(contains(r'$')));
     }
   }
 
@@ -472,6 +603,18 @@ build() => Contai^ner(child: Container());
         DartAssistKind.flutterWrapGeneric.message,
         DartAssistKind.flutterRemoveWidget.message,
       ]),
+    );
+  }
+
+  LegacySnippetTextEdit _extractLegacySnippetTextEdit(
+    Either4<AnnotatedTextEdit, LegacySnippetTextEdit, SnippetTextEdit, TextEdit>
+    edit,
+  ) {
+    return edit.map(
+      (e) => throw 'Expected LegacySnippetTextEdit, got AnnotatedTextEdit',
+      (e) => e,
+      (e) => throw 'Expected LegacySnippetTextEdit, got SnippetTextEdit',
+      (e) => throw 'Expected LegacySnippetTextEdit, got TextEdit',
     );
   }
 }